remove dead indexing code (#1806)
